+++ app/cdap/components/WrangleHome/index.tsx
@@ -1,13 +1,13 @@
 import { Box } from '@material-ui/core';
 import ConnectorTypesComponent from 'components/ConnectorTypes';
-import ConnectionTabsCaller from 'components/Dataset/ConnectionTabsCaller';
+import Dataset from 'components/Dataset/index';
 import * as React from 'react';
 
 const WrangleHome: React.FC = () => {
   return (
     <Box>
       {/* <ConnectorTypesComponent /> */}
-      <ConnectionTabsCaller />
+      <Dataset />
     </Box>
   );
 };
